address a few issues in the chrome lighthouse audit (#12)
a few issues got flagged in the lighthouse audit, and are addressed in this commit:

- lack of "lang" attribute in the html element
- lack of "alt" attribute on images

We've added a default `alt=''` attribute for the images that don't explicitly have `alt` tags, since for some screen readers, the default behavior is to read the `src` attribute of the element, and we don't want that.

- lack of "meta description"
- insecure external links

Most modern browsers support "noopener", though for completeness it's still common to combine it with "noreferrer" for support in older browsers. This addition is to prevent clickjacking vulnerabilities.