add basic cypress end-to-end tests (#20)
It's helpful to be able to assert that the pages are rendering fine for the users, even if some of the underlying code implementation changes. With cypress, we can use this in CI to test whether we've broken any pages, which _shouldn't_ happen with gatsby, but always nice to have an extra check.

The additional npm scripts are mostly for local dev, since the github action workflow is what we'll actually use in CI.

The tests in this commit are super simple, and eventually we probably want to test for accessibility issues, and more complex button clicking and form-filling. The weird empty-ish files in `cypress/plugins` and `cypress/support` would be used for that, so I added them here just we already have them.